Add active status to volunteer
Because this patchset involves an API change, many files were
touched.

The ManageVolunteerPanel class has not yet had the appropriate
functionality implemented to deal with this change; it and its unit
tests have been altered in order to compile/pass, but a new patch is
necessary to implement its functionality.

This change relates to issue #18.