Redo ManageListFrame resizing code
An early attempt at this patch set the minimum size to the value of
"getMinimumSize()". However, this caused some unfortunate problems
with ManageVolunteersFrame; the JTextArea tended to shrink down to a
single row when the grame was resized to its minimum size. Using
"getPreferredSize()" solves this problem.

This commit relates to issue #52.